Rename quantaq_cli to quantaq_py
