+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
-name = "quantaq-cli"
+name = "quantaq-py"
 version = "1.0.0rc0"
-description = "The QuantAQ CLI"
+description = "The QuantAQ Python Toolkit and CLI"
 authors = ["David H Hagan <david.hagan@quant-aq.com>"]
 license = "Apache-2.0"
 readme = "README.md"
@@ -33,7 +33,7 @@ sphinx-rtd-theme = "^2.0"
 setuptools = "^65.6.3"
 
 [tool.poetry.scripts]
-quantaq-cli = "quantaq_cli.console:main"
+quantaq-py = "quantaq_py.console:main"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
